requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
email-validator==2.1.0
phonenumbers==8.13.19
requests==2.31.0
Pillow==10.0.0
psycopg2-binary==2.9.9  # For PostgreSQL, optional